Statement 생성, Connection, Logging 책임을 SimpleDb로 위임
