--- src/main/java/com/ll/DataSource.java
@@ -1,9 +1,0 @@
-package com.ll;
-
-
-public interface DataSource {
-    void closeConnection();
-    void startTransaction();
-    void rollback();
-    void commit();
-}
